refactor: move submit_extrinsic_with_wallet in a common file
